requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
drf-yasg2==1.19.4
fonttools==4.39.4
fpdf==1.7.2
fpdf2==2.7.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyaml==23.5.9
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
reportlab==4.0.4
requests==2.30.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
